Adding auth to all API endpoints.

+++ backend/categories/schema.py
@@ -1,6 +1,6 @@
 import graphene
 from graphene_django import DjangoObjectType
-# from gqlauth.validators import validate_username, validate_password, validate_user_is_authenticated
+from uauth.validators import validate_user_is_authenticated, validate_user_is_admin
 from django.contrib.auth import authenticate, login, logout
 from django.contrib.auth.models import User
 
@@ -14,6 +14,7 @@ class Query(graphene.ObjectType):
     all_categories = graphene.List(CategoryType)
 
     def resolve_all_categories(self, info, **kwargs):
+        validate_user_is_authenticated(info.context.user)
         return Category.objects.all()
 
 # ------------------- MUTATIONS -------------------
@@ -26,6 +27,7 @@ class AddCategory(graphene.Mutation):
         description = graphene.String(required=True)
 
     def mutate(self, info, name, description):
+        validate_user_is_admin(info.context.user)
         try:
             newCategory = Category(name=name, description=description)
             newCategory.save()
@@ -42,6 +44,7 @@ class RemoveCategory(graphene.Mutation):
         id = graphene.Int(required=True)
 
     def mutate(self, info, id):
+        validate_user_is_admin(info.context.user)
         try:
             category = Category.objects.get(pk=id)
             category.delete()
@@ -60,6 +63,7 @@ class UpdateCategory(graphene.Mutation):
         description = graphene.String(required=True)
 
     def mutate(self, info, id, name, description):
+        validate_user_is_admin(info.context.user)
         try:
             category = Category.objects.get(pk=id)
             category.name = name
